Time: 3 ms (36.94%), Space: 6 MB (62.52%) - LeetHub

+++ 670-maximum-swap/670-maximum-swap.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    int maximumSwap(int num)
+    int maximumSwap(long long num)
     {
         string numString = to_string(num);
         int n = numString.length();
